fix: separated the Nav Menu to its own component
