+++ packages/templates/nextjs-marketing-demo/src/components/layout/Footer.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import React from 'react';
 import { Menu } from 'antd';
 import styles from './footer.module.css';
