+++ packages/templates/nextjs-marketing-demo/src/components/layout/nav.module.css
@@ -1,0 +1,4 @@
+ul.menu {
+  border-bottom: 1px solid transparent;
+}
+
